Move Stateful macro into BytesStateful in openvm-circuit-primitives-derive
