+++ benchmarks/Cargo.toml
@@ -10,7 +10,7 @@ license.workspace = true
 [dependencies]
 openvm-build.workspace = true
 openvm-circuit.workspace = true
-openvm-circuit-derive.workspace = true
+openvm-circuit-primitives-derive.workspace = true
 openvm-sdk.workspace = true
 openvm-stark-backend.workspace = true
 openvm-stark-sdk.workspace = true
